// src/core/authorization/rol.enum.ts
export enum RolEnum {
  // USUARIOS
  SUPER_ADMIN = 'SUPER_ADMIN',
  ADMINISTRADOR = 'ADMINISTRADOR',
  CONSULTA = 'CONSULTA',
  // ACTORES
  COMITE_DOCENTE_CARGA = 'COMITE_DOCENTE_CARGA',
  COMITE_DOCENTE_VERIFICADOR = 'COMITE_DOCENTE_VERIFICADOR',
  // COMITE_DOCENTE_CORRECCION = 'COMITE_DOCENTE_CORRECCION',
  COMITE_DEPARTAMENTAL = 'COMITE_DEPARTAMENTAL',
  TECNICO_SIE_DEPARTAMENTAL = 'TECNICO_SIE_DEPARTAMENTAL',
  TECNICO_SIE_DISTRITO = 'TECNICO_SIE_DISTRITO',
  DIRECTOR = 'DIRECTOR',
  // ESTUDIANTE
  ESTUDIANTE = 'ESTUDIANTE',
}